+++ packages/rum/src/boot/startRecording.spec.ts
@@ -1,4 +1,4 @@
-import { HttpRequest } from '@datadog/browser-core'
+import { HttpRequest, DefaultPrivacyLevel } from '@datadog/browser-core'
 import { LifeCycle, LifeCycleEventType } from '@datadog/browser-rum-core'
 import { inflate } from 'pako'
 import { createRumSessionMock, RumSessionMock } from '../../../rum-core/test/mockRumSession'
@@ -45,13 +45,14 @@ describe('startRecording', () => {
             },
             view: {
               id: viewId,
-              referrer: '',
-              url: 'http://example.org',
             },
           }
         },
       })
       .withSession(session)
+      .withConfiguration({
+        defaultPrivacyLevel: DefaultPrivacyLevel.ALLOW,
+      })
       .beforeBuild(({ lifeCycle, applicationId, configuration, parentContexts, session }) => {
         ;({ stop: stopRecording } = startRecording(lifeCycle, applicationId, configuration, session, parentContexts))
         return { stop: stopRecording }