+++ packages/rum/src/domain/record/serializationUtils.ts
@@ -1,14 +1,13 @@
 import { buildUrl } from '@datadog/browser-core'
 import { CENSORED_STRING_MARK, NodePrivacyLevel } from '../../constants'
 import { shouldMaskNode } from './privacy'
-import { SerializedNodeWithId } from './types'
 
-export interface NodeWithSerializedNode extends Node {
-  __sn: SerializedNodeWithId
-}
+export type NodeWithSerializedNode = Node & { s: 'Node with serialized node' }
+
+const serializedNodeIds = new WeakMap<Node, number>()
 
 export function hasSerializedNode(node: Node): node is NodeWithSerializedNode {
-  return '__sn' in node
+  return serializedNodeIds.has(node)
 }
 
 export function nodeAndAncestorsHaveSerializedNode(node: Node): node is NodeWithSerializedNode {
@@ -25,11 +24,11 @@ export function nodeAndAncestorsHaveSerializedNode(node: Node): node is NodeWith
 export function getSerializedNodeId(node: NodeWithSerializedNode): number
 export function getSerializedNodeId(node: Node): number | undefined
 export function getSerializedNodeId(node: Node) {
-  return hasSerializedNode(node) ? node.__sn.id : undefined
+  return serializedNodeIds.get(node)
 }
 
-export function setSerializedNode(node: Node, serializeNode: SerializedNodeWithId) {
-  ;(node as Partial<NodeWithSerializedNode>).__sn = serializeNode
+export function setSerializedNodeId(node: Node, serializeNodeId: number) {
+  serializedNodeIds.set(node, serializeNodeId)
 }
 
 const URL_IN_CSS_REF = /url\((?:(')([^']*)'|(")([^"]*)"|([^)]*))\)/gm